@@ -14,48 +14,45 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "5.0.1",
-    "@angular/cdk": "^5.0.0-rc0",
-    "@angular/common": "5.0.1",
-    "@angular/compiler": "5.0.1",
-    "@angular/core": "5.0.1",
-    "@angular/flex-layout": "^2.0.0-beta.10-4905443",
-    "@angular/forms": "5.0.1",
-    "@angular/http": "5.0.1",
-    "@angular/material": "^5.0.0-rc0",
-    "@angular/material-moment-adapter": "^5.0.0-rc.1",
-    "@angular/platform-browser": "5.0.1",
-    "@angular/platform-browser-dynamic": "5.0.1",
-    "@angular/router": "5.0.1",
-    "@ngx-translate/core": "^8.0.0",
-    "@ngx-translate/http-loader": "2.0.0",
-    "angular2-jwt": "0.2.3",
-    "classlist.js": "^1.1.20150312",
-    "core-js": "^2.5.1",
-    "crypto-js": "^3.1.9-1",
-    "file-saver": "^1.3.3",
-    "hammerjs": "^2.0.8",
-    "intl": "^1.2.5",
-    "js-marker-clusterer": "^1.0.0",
-    "moment": "^2.19.2",
-    "ng2-file-upload": "^1.2.1",
-    "ngx-color-picker": "^5.0.1",
-    "node-sass": "^4.6.1",
-    "rxjs": "5.5.2",
+    "@angular/animations": "6.0.9",
+    "@angular/cdk": "^6.4.0",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "6.0.9",
+    "@angular/core": "6.0.9",
+    "@angular/flex-layout": "6.0.0-beta.16",
+    "@angular/forms": "6.0.9",
+    "@angular/http": "6.0.9",
+    "@angular/material": "6.4.0",
+    "@angular/material-moment-adapter": "6.4.0",
+    "@angular/platform-browser": "6.0.9",
+    "@angular/platform-browser-dynamic": "6.0.9",
+    "@angular/router": "6.0.9",
+    "@auth0/angular-jwt": "2.0.0",
+    "@ngx-translate/core": "10.0.2",
+    "@ngx-translate/http-loader": "3.0.1",
+    "classlist.js": "1.1.20150312",
+    "core-js": "2.5.7",
+    "crypto-js": "3.1.9-1",
+    "file-saver": "1.3.8",
+    "hammerjs": "2.0.8",
+    "intl": "1.2.5",
+    "moment": "2.22.2",
+    "ng2-file-upload": "^1.3.0",
+    "node-sass": "4.9.2",
+    "rxjs": "6.2.2",
     "ts-helpers": "1.1.2",
-    "underscore": "^1.8.3",
+    "underscore": "^1.9.1",
     "web-animations-js": "2.3.1",
-    "zone.js": "0.8.18"
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.5.0",
-    "@angular/compiler-cli": "5.0.1",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.9",
     "@types/file-saver": "1.3.0",
-    "@types/jasmine": "2.8.0",
+    "@types/jasmine": "2.8.8",
     "@types/node": "8.0.51",
-    "codelyzer": "4.0.1",
-    "jasmine-core": "2.8.0",
+    "codelyzer": "4.4.2",
+    "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
@@ -63,10 +60,10 @@
     "karma-coverage-istanbul-reporter": "1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "markerwithlabel": "^2.0.1",
     "protractor": "~5.2.0",
     "ts-node": "3.3.0",
     "tslint": "5.8.0",
-    "typescript": "2.5.2"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
