@@ -1,15 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
+import { of } from 'rxjs';
+
 
 import { StorageService } from '../../core/storage.service';
 import { AuthorizationService } from '../authorization.service';
 import { BaseTranslationComponent } from '../../core/base-translation-component';
-import { LoginModel } from '../models/login-model';
+import { catchError } from 'rxjs/operators';
 
 /**
  * Represents the login component.
@@ -41,7 +41,7 @@ export class LoginComponent extends BaseTranslationComponent implements OnInit {
    * Initializes a new instance of the LoginComponent class.
    * @constructor
    * @param {Router} router The current router.
-   * @param {Http} http The current http service.
+   * @param {HttpClient} http The current http service.
    * @param {FormBuilder} formBuilder The angular form builder.
    * @param {AuthorizationService} authorizationService The application authorization service.
    * @param {StorageService} storage The application storage service.
@@ -49,7 +49,7 @@ export class LoginComponent extends BaseTranslationComponent implements OnInit {
    */
   constructor(
     private router: Router,
-    private http: Http,
+    private http: HttpClient,
     private formBuilder: FormBuilder,
     private authorizationService: AuthorizationService,
     storage: StorageService,
@@ -82,19 +82,19 @@ export class LoginComponent extends BaseTranslationComponent implements OnInit {
       return;
     }
 
-    this.authorizationService.login(this.loginForm.value)
-      .catch(() => {
+    this.authorizationService.login(this.loginForm.value).pipe(
+      catchError((error) => {
         this.authenticationFailed = true;
-        return Observable.of({});
+        return of({});
       })
-      .subscribe(response => {
-        if (this.authenticationFailed) {
-          return;
-        }
+    ).subscribe(response => {
+      if (this.authenticationFailed) {
+        return;
+      }
 
-        const redirect = this.authorizationService.redirectUrl || '';
-        this.router.navigate([redirect]);
-      });
+      const redirect = this.authorizationService.redirectUrl || '';
+      this.router.navigate([redirect]);
+    });
   }
 
 }
