@@ -3,15 +3,13 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { TranslateModule } from '@ngx-translate/core';
-import { Http } from '@angular/http';
-import { AuthHttp } from 'angular2-jwt';
+import { HttpClient } from '@angular/common/http';
 
 import { LoginComponent } from './login/login.component';
 import { LogoutComponent } from './logout/logout.component';
 import { ForgotPasswordComponent } from './forgot-password/forgot-password.component';
 import { ForbiddenComponent } from './forbidden/forbidden.component';
 import { AuthorizationService } from './authorization.service';
-import { authorizationHttpFactory } from './authorization-http';
 import { HasRightDirective } from './has-right.directive';
 import { USERS_GUARD_PROVIDERS } from './guards';
 import { MatButtonModule, MatCardModule, MatInputModule, MatIconModule } from '@angular/material';
@@ -51,12 +49,8 @@ export class AuthenticationModule {
       ngModule: AuthenticationModule,
       providers: [
         AuthorizationService,
-        USERS_GUARD_PROVIDERS,
-        {
-          provide: AuthHttp,
-          useFactory: authorizationHttpFactory,
-          deps: [AuthorizationService, Http]
-        }]
+        USERS_GUARD_PROVIDERS
+      ]
     };
   };
 }
