@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import 'rxjs/add/operator/toPromise';
+
 
 import { AuthorizationService } from '../authorization.service';
 
