@@ -1,12 +1,15 @@
-import { Http, Response } from '@angular/http';
+import { Response } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 
 import { StorageService } from './storage.service';
 import { environment } from '../../environments/environment';
+import { map } from 'rxjs/operators';
 
 export const showTranslationKey = 'app.stc';
 
-declare class TranslateResponse extends Response {
-    json(): Array<{ key: string, value: string }>;
+declare class TranslateModel {
+    key: string;
+    value: string;
 }
 
 /**
@@ -18,12 +21,12 @@ export class ApiTranslateLoader {
     /**
      * Initializes anew instance of the ApiTranslateLoader class.
      * @constructor
-     * @param {Http} http The angular http service.
+     * @param {HttpClient} http The angular http service.
      * @param {StorageService} storage The application storage service.
      * @param {string} url The url translations.
      */
     constructor(
-        private http: Http,
+        private http: HttpClient,
         private storage: StorageService,
         private url: string
     ) {
@@ -37,24 +40,26 @@ export class ApiTranslateLoader {
     getTranslation(lang: string) {
         const show = this.storage.getItem<boolean>(showTranslationKey) || false;
 
-        return this.http.get(`${this.url}/${lang}/display?timestamp=${Date.now()}`).map((data: TranslateResponse) => {
-            const result = new Map<string, string>();
-            data.json().forEach(element => {
-                result[element.key] = show ? element.key + ' (' + element.value + ')' : element.value;
-            });
+        return this.http.get(`${this.url}/${lang}/display?timestamp=${Date.now()}`).pipe(
+            map((data: TranslateModel[]) => {
+                const result = new Map<string, string>();
+                data.forEach(element => {
+                    result[element.key] = show ? element.key + ' (' + element.value + ')' : element.value;
+                });
 
-            return result;
-        });
+                return result;
+            })
+        );
     }
 }
 
 /**
  * Function which represents the translate loader factory.
  * @method
- * @param {Http} http The angular http service.
+ * @param {HttpClient} http The angular http service.
  * @param {StorageService} storage The application storage service.
  * @return {ApiTranslateLoader}
  */
-export function apiTranslateLoaderFactory(http: Http, storage: StorageService) {
+export function apiTranslateLoaderFactory(http: HttpClient, storage: StorageService) {
     return new ApiTranslateLoader(http, storage, `${environment.apiUrlBase}api/languages`);
 }
